requirements.txt
gitdb==4.0.11
GitPython==3.1.43
install-jdk==1.1.0
smmap==5.0.1
